Added async/promise functionality
Wow what a ride. I got asynchronous functionality working. First, `sendSlackMessage` will initiate `web.im.open` for the user the message is being sent to. Then, using `.then()` the message payload is sent to the user.

Future functionality:
I'd like to add a `console.log()` event after the message is sent that lists the user name and the message sent and the time sent, etc.